requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
channels==4.1.0
distro==1.9.0
Django==5.1
django-nextjs==3.2.0
djangorestframework==3.15.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
multidict==6.1.0
nodejs==0.1.1
npm==0.1.1
openai==1.47.0
optional-django==0.1.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
websockets==13.0.1
yarl==1.11.1
